MPDDPM_p/requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2023.12.0
idna==3.6
Jinja2==3.1.3
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
Pillow==10.2.0
requests==2.31.0
sympy==1.12
torch==2.1.1
torchaudio==2.1.1
torchsummary==1.5.1
torchvision==0.16.1
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.1.0
